refactor(catalog-extension-api): move it from main to api package
so it can be used by preload, main and renderer

related to #14361
and #15485
