+++ packages/main/src/plugin/extension/catalog/extensions-catalog.ts
@@ -23,12 +23,9 @@ import { inject, injectable } from 'inversify';
 
 import { ApiSenderType } from '/@/plugin/api.js';
 import { Certificates } from '/@/plugin/certificates.js';
-import type {
-  CatalogExtension,
-  CatalogFetchableExtension,
-} from '/@/plugin/extension/catalog/extensions-catalog-api.js';
 import { Proxy } from '/@/plugin/proxy.js';
 import { type IConfigurationNode, IConfigurationRegistry } from '/@api/configuration/models.js';
+import { CatalogExtension, CatalogFetchableExtension } from '/@api/extension-catalog/extensions-catalog-api.js';
 import product from '/@product.json' with { type: 'json' };
 
 import { ExtensionsCatalogSettings } from './extensions-catalog-settings.js';
