+++ packages/main/src/plugin/install/extension-installer.spec.ts
@@ -23,9 +23,9 @@ import type { IpcMain, IpcMainEvent } from 'electron';
 import { beforeEach, expect, test, vi } from 'vitest';
 
 import type { ExtensionsCatalog } from '/@/plugin/extension/catalog/extensions-catalog.js';
-import type { CatalogFetchableExtension } from '/@/plugin/extension/catalog/extensions-catalog-api.js';
 import type { AnalyzedExtension } from '/@/plugin/extension/extension-analyzer.js';
 import type { ExtensionLoader } from '/@/plugin/extension/extension-loader.js';
+import type { CatalogFetchableExtension } from '/@api/extension-catalog/extensions-catalog-api.js';
 import type { ExtensionInfo } from '/@api/extension-info.js';
 
 import type { ApiSenderType } from '../api.js';
