+++ packages/main/src/plugin/recommendations/recommendations-registry.spec.ts
@@ -20,10 +20,10 @@ import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
 
 import type { ConfigurationRegistry } from '/@/plugin/configuration-registry.js';
 import type { ExtensionsCatalog } from '/@/plugin/extension/catalog/extensions-catalog.js';
-import type { CatalogFetchableExtension } from '/@/plugin/extension/catalog/extensions-catalog-api.js';
 import type { ExtensionLoader } from '/@/plugin/extension/extension-loader.js';
 import type { Featured } from '/@/plugin/featured/featured.js';
 import type { FeaturedExtension } from '/@/plugin/featured/featured-api.js';
+import type { CatalogFetchableExtension } from '/@api/extension-catalog/extensions-catalog-api.js';
 import type { ExtensionInfo } from '/@api/extension-info.js';
 
 import { RecommendationsRegistry } from './recommendations-registry.js';
