+++ packages/renderer/src/lib/extensions/EmbeddableCatalogExtensionList.spec.ts
@@ -24,8 +24,8 @@ import { beforeAll, beforeEach, expect, test, vi } from 'vitest';
 import type { CombinedExtensionInfoUI } from '/@/stores/all-installed-extensions';
 import { catalogExtensionInfos } from '/@/stores/catalog-extensions';
 import { extensionInfos } from '/@/stores/extensions';
+import type { CatalogExtension } from '/@api/extension-catalog/extensions-catalog-api';
 
-import type { CatalogExtension } from '../../../../main/src/plugin/extension/catalog/extensions-catalog-api';
 import EmbeddableCatalogExtensionList from './EmbeddableCatalogExtensionList.svelte';
 
 beforeAll(() => {
@@ -36,7 +36,7 @@ beforeEach(() => {
   vi.resetAllMocks();
 });
 
-export const aFakeExtension: CatalogExtension = {
+const aFakeExtension: CatalogExtension = {
   id: 'idAInstalled',
   publisherName: 'FooPublisher',
   shortDescription: 'this is short A',
