+++ packages/renderer/src/lib/extensions/ExtensionList.spec.ts
@@ -24,8 +24,8 @@ import { beforeEach, expect, test, vi } from 'vitest';
 import { type CombinedExtensionInfoUI } from '/@/stores/all-installed-extensions';
 import { catalogExtensionInfos } from '/@/stores/catalog-extensions';
 import { extensionInfos } from '/@/stores/extensions';
+import type { CatalogExtension } from '/@api/extension-catalog/extensions-catalog-api';
 
-import type { CatalogExtension } from '../../../../main/src/plugin/extension/catalog/extensions-catalog-api';
 import ExtensionList from './ExtensionList.svelte';
 
 beforeEach(() => {
