+++ packages/renderer/src/lib/extensions/extensions-utils.ts
@@ -17,8 +17,8 @@
  ***********************************************************************/
 
 import type { CombinedExtensionInfoUI } from '/@/stores/all-installed-extensions';
+import type { CatalogExtension } from '/@api/extension-catalog/extensions-catalog-api';
 
-import type { CatalogExtension } from '../../../../main/src/plugin/extension/catalog/extensions-catalog-api';
 import type { FeaturedExtension } from '../../../../main/src/plugin/featured/featured-api';
 import type { CatalogExtensionInfoUI } from './catalog-extension-info-ui';
 import type { ExtensionDetailsUI } from './extension-details-ui';
