+++ packages/renderer/src/stores/catalog-extensions.spec.ts
@@ -22,7 +22,8 @@ import { get } from 'svelte/store';
 import type { Mock } from 'vitest';
 import { beforeAll, expect, test, vi } from 'vitest';
 
-import type { CatalogExtension } from '../../../main/src/plugin/extension/catalog/extensions-catalog-api';
+import type { CatalogExtension } from '/@api/extension-catalog/extensions-catalog-api';
+
 import {
   catalogExtensionEventStore,
   catalogExtensionEventStoreInfo,
